Block 7 — MockPlatform + headless app driver + golden screens
Add a scripted MockPlatform and a poll-based AppDriver (app::testing,
feature `testing`) so every built-in app is testable in CI without
hardware or a display, plus golden-screen integration tests.

- app/src/testing/mock.rs — MockPlatform: EventQueue-backed input/system/
  hexpansion, in-memory BTree storage + config, canned HTTP/TCP, and a
  recording DisplayManager.
- app/src/testing/driver.rs — AppDriver drives an app's run() loop with a
  no-op waker, advancing the embassy-time mock clock 1 ms per settle poll.
- app/tests/golden.rs + app/tests/golden/<app>/<scenario>.json — golden
  tests for all 10 apps (launch + one interaction); regenerate with
  UPDATE_GOLDEN=1 (idempotent).
- justfile: `just test` runs `cargo test -p app --features testing`.
- power.rs: fix a u16 overflow in battery_percent() (panics in debug,
  wraps wrong in release) that the PowerInfo golden test exposed.
- AGENTS.md: document the testing module + golden workflow; drop the now-
  done "Low 2" remaining-work item.

+++ app/src/platform/power.rs
@@ -22,8 +22,10 @@ pub struct PowerStatus {
 
 impl PowerStatus {
   pub fn battery_percent(&self) -> u8 {
-    // Li-ion nominal range ~3.0V–4.2V, map 0–100%
-    let mv = self.vbat_mv.clamp(3000, 4200);
+    // Li-ion nominal range ~3.0V–4.2V, map 0–100%.
+    // Compute in `u32` — `(4200 - 3000) * 100` overflows `u16` and would panic
+    // in debug builds (or wrap to a wrong value in release).
+    let mv = self.vbat_mv.clamp(3000, 4200) as u32;
     ((mv - 3000) * 100 / (4200 - 3000)) as u8
   }
 }
